Add files via upload
Energy supply for economic activities has been a major problem in Nigeria such that it does appear that solution 
to stable, reliable power supply is not yet in sight. Consequently, consumers (industrial, residential and 
commercial) have all resorted to self-generation to complement whatever supply (if any) that comes from the very 
unreliable national grid. For instance, the available power supply to the entire nation fluctuates between 3000MW 
and 3500MW. A study of monthly energy generation in Dangote Cement Plant for a typical year was carried out. 
This was compared with the equivalent monthly generation from the entire Nigeria’s grid system. The results 
show that in 2013, at the early stage of commencement of production of operation in the plant), the total energy 
generated by the Dangote Power Plant was 406762.33 MWH while the total energy consumed by the company 
was 401903.72 MWh. Dangote’s total energy generation in 2013 represents 1.4% of Nigeria’s total generation 
from grid in the same year. The total distribution loss was 4858.61 MWH (which is 1.2% of the total energy 
generated by the company).Comparatively, the total generation from national grid in 2013 was 29628699. This 
study further shows that Dangote Cement, Ibese produces enough energy for its use and could even sell any excess 
if encouraged.