Add files via upload
Experts are varied in estimating the amount of power needed for national development. One expert 
estimated 297,900 MW by the year 2030 using a 13 percent Gross Domestic Product (GDP). The Power 
Sector Road Map’s aspiration is 40,000 MW by the year 2020 while the preferred Vision 20:2020 target is 
40GW (40,000 MW) available capacity. Vision 20:2020 further proposed that hydro sources contribute 10% 
of this value; thermal, 80%; coal, 6% and renewables, 4%. However, as at December 2017, Nigeria’s total
installed generating capacity was 12,324.40 MW. In order to assess the fundamentals and proffer solutions 
for the improvement of power generation to meet popular expectations, this work analyses the installed 
capacities of the nation from 1986 – 1995, on the one hand and from 2007 – 2016, on the other hand using 
graphical illustrations and tables. The gap between the two time phases was deliberate to create some effect. 
The results show that over the years, there has been an apparent, non – challance towards systematic 
development of Nigeria’s power sector. To put a check to this deteriorating condition / trend, useful 
suggestions have been made.