Add files via upload
Nigeria is facing an acute shortage of 
electric power that drives economic activities. 
Consequently, all sectors of the economy are facing 
a retrogression in productivity. Besides, even much 
of the power made available either from the grid or 
self-generation is wasted due to poor installation 
design and use of inefficient devices and equipment 
such as filament bulbs, and fluorescent lamps 
(among others). This work examines the present 
common practice of electrical services that wastes 
energy with the viewing to advancing the best 
practice that minimises overall energy / power loss. A 
three-bedroom flat and an eatery in Nigeria were 
used as a case study. Load audit, analysis and 
calculations were made to assist in making vital 
decisions as to the types and sizes of luminaires, 
cables and protective devices (among others) to be 
used in accordance with known standards and 
regulations. The results show that 71.8% and 69.1% 
respectively of energy is saved for lighting in three-
bedroom flat and eatery; 37.4% and 30.8% in other 
loads used in three bedroom and eatery respectively. 
All this is as a result of use of frugal design and 
energy saving – saving devices. This further implies 
a saving in absolute cost of energy consumed in a 
particular situation.