Add files via upload
Solar Energy is the prime important 
source of energy, and it has continued to gain 
popularity globally. As of 2018, about 486 GW of solar 
PV was installed worldwide. One of the key 
requirements for socio-economic improvement in any 
nation of the world is the provision of dependable 
electricity supply systems. Recently, there is a massive 
growth in access to solar electricity in several Africa 
countries, notably South Africa, Egypt, Morocco, and 
Algeria. As a result, it decreases the global population 
without access to electricity with an appreciable value. 
This paper review the present state of solar energy 
capacity in the world also identifies vital approaches of 
improving their functionality, reliability, and 
affordability as well as the essential method that 
policymakers may implement in the future. The result 
shows that there is a recurrent growth in solar energy 
annually and as at the end of 2018 Asia with the largest 
capacity of 56.58%, while Central America and the 
Caribbean have the lowest installed capacity of 0.36%. 
As the request for an uninterrupted supply increases in 
different regions, progressive builders are embracing 
solar photovoltaics (PV) renewable energy as an option 
for their customers.